Use downloaded track for memories music
