+++ mobile/apps/photos/lib/models/memories/memory_music_track.dart
@@ -1,12 +1,14 @@
 class MemoryMusicTrack {
   final String id;
-  final String assetPath;
+  final String url;
+  final String cacheFileName;
   final String? title;
   final String? artist;
 
   const MemoryMusicTrack({
     required this.id,
-    required this.assetPath,
+    required this.url,
+    required this.cacheFileName,
     this.title,
     this.artist,
   });
