--- mobile/apps/photos/lib/services/memories/bundled_memory_music_tracks.dart
@@ -1,22 +1,0 @@
-import "package:photos/models/memories/memory_music_track.dart";
-
-const bundledMemoryMusicTracks = <MemoryMusicTrack>[
-  MemoryMusicTrack(
-    id: "puddles-instrumental-segment-1",
-    assetPath: "assets/memory-music/puddles-instrumental-01.mp3",
-    title: "Puddles (Zero 7 Remix)",
-    artist: "Not for Radio",
-  ),
-  MemoryMusicTrack(
-    id: "puddles-instrumental-segment-2",
-    assetPath: "assets/memory-music/puddles-instrumental-02.mp3",
-    title: "Puddles (Zero 7 Remix)",
-    artist: "Not for Radio",
-  ),
-  MemoryMusicTrack(
-    id: "puddles-instrumental-segment-3",
-    assetPath: "assets/memory-music/puddles-instrumental-03.mp3",
-    title: "Puddles (Zero 7 Remix)",
-    artist: "Not for Radio",
-  ),
-];
