+++ mobile/apps/photos/lib/services/memories/memory_music_tracks.dart
@@ -1,0 +1,11 @@
+import "package:photos/models/memories/memory_music_track.dart";
+
+const memoryMusicTracks = <MemoryMusicTrack>[
+  MemoryMusicTrack(
+    id: "then",
+    url: "https://music.ente.com/then.mp3",
+    cacheFileName: "memory-music-then.mp3",
+    title: "Then",
+    artist: "Vishnu Mohandas",
+  ),
+];
