+++ mobile/apps/photos/pubspec.yaml
@@ -282,7 +282,6 @@ flutter:
   assets:
     - assets/
     - assets/rituals/
-    - assets/memory-music/
     - assets/rewind_badges/
     - assets/video-editor/
     - assets/image-editor/
